Change win probability to int; fixed CardValidator in context with win probability; Fixed js

+++ src/main/java/de/lemonpie/beddomischer/http/websocket/listener/PlayerCallbackListener.java
@@ -59,7 +59,7 @@ public class PlayerCallbackListener implements PlayerListener {
 	}
 
 	@Override
-	public void winProbabilityDidChange(Player player, double value) {
+	public void winProbabilityDidChange(Player player, int value) {
 		// Handled in WinProbabilityPlayerListener
 	}
 }
